+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.5.6 ##
+
+* Fix external zlib issue caused by 1.2.7 being removed from source by upgrading to version 1.2.8 (thanks @hughevans). This unfortunately breaks previously installed versions, however this is fixed by the provider changes
+* Improve `php_source` provider for `php_version` so that a version of PHP is tested to confirm it is working. Failing this check will then trigger a recompile. As an example this will fix the zlib library version changing.
+
+## 0.5.5 ##
+
+* Add additional PHP versions, thanks @webflo
+* Add new `php_source` extension provider which can install php extensions bundled with the PHP source code
+
 ## 0.5.4 ##
 
 * Fix config issue with xdebug extension